80102570 regen; needs work

+++ src/ovl2/ovl2_6.h
@@ -315,7 +315,7 @@ struct CollisionState {
     struct Normal *unk38;
     u32 (*unk3C)(void);
     u8 (*unk40)(struct CollisionTriangle *a0, struct Normal *a1, struct Normal *a2, struct Normal *a3);
-    u32 (*unk44)(struct Normal *a0, s32 arg1);
+    u8 (*unk44)(struct Normal *a0, s32 arg1);
     u16 unk48;
     u16 unk4A;
     u16 unk4C;
@@ -328,4 +328,5 @@ struct CollisionState {
 #define NO_SHADOW           (1 << 2)
 #define NON_SOLID           (1 << 3)
 #define DOUBLE_SIDED_NORMAL (FORWARD_NORMAL | BACKWARD_NORMAL)
+
